moved and organized the telemetry files and updated the example env file to be more descriptive
