+++ .env.example
@@ -4,9 +4,10 @@
 
 # ============================================================================
 # DEVELOPMENT FLAGS
+# Recomend not changing these unless you know what you're doing they are set by the launch.json normally
 # ============================================================================
-IS_DEV=true
-CLINE_ENVIRONMENT=local
+# IS_DEV=true
+# CLINE_ENVIRONMENT=local
 
 # ============================================================================
 # POSTHOG TELEMETRY (Existing)
@@ -28,7 +29,7 @@ JITSU_HOST=https://jitsu.cline.bot
 # Control which telemetry providers are active
 JITSU_ENABLED=true                      # Enable Jitsu telemetry (default: true)
 POSTHOG_TELEMETRY_ENABLED=true          # Enable PostHog telemetry (default: true)
-                                        # Set to false to disable PostHog completely
+                                        # Set to false to disable Telemetry completely
 
 # ============================================================================
 # OPTIONAL DEVELOPMENT SETTINGS
